페이징, 파일첨부 작업

+++ culture/src/main/java/com/ije/controller/CultureController.java
@@ -3,12 +3,15 @@ package com.ije.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.ije.domain.Criteria;
 import com.ije.domain.CultureVO;
+import com.ije.domain.PageVO;
 import com.ije.service.CultureService;
 
 import lombok.RequiredArgsConstructor;
@@ -22,14 +25,21 @@ public class CultureController {
 
 	private final CultureService service; 
 	
-	@GetMapping("/list")
+	/*@GetMapping("/list")
 	public void list(Model d) {
 		log.info("전체 목록 호출..................................");
 		d.addAttribute("list", service.getList()); 
-	}
+	}*/
 	
+	@GetMapping("/list")
+	public void list(Criteria cri, Model d) {
+		log.info("페이징 호출..............................................");
+		d.addAttribute("list", service.getListPaging(cri)); 
+		d.addAttribute("page", new PageVO(cri, service.getCount(cri)));
+	}
+		
 	@GetMapping("/get")
-	public void get(@RequestParam("cno") Long cno, Model d) {
+	public void get(@RequestParam("cno") Long cno, @ModelAttribute("cri") Criteria cri, Model d) {
 		log.info("상세보기 호출.........................................");
 		log.info("cno : " + cno);
 		log.info("...........................................");
@@ -51,7 +61,7 @@ public class CultureController {
 	}
 	
 	@GetMapping("/modify")
-	public void modify(@RequestParam("cno") Long cno, Model d) {
+	public void modify(@RequestParam("cno") Long cno, @ModelAttribute("cri") Criteria cri, Model d) {
 		log.info("수정폼 호출...........................................");
 		log.info(service.get(cno));
 		log.info("...........................................");
@@ -59,20 +69,28 @@ public class CultureController {
 	}
 	
 	@PostMapping("/modify")
-	public String modify(CultureVO upt, RedirectAttributes rttr) {
+	public String modify(CultureVO upt, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr) {
 		log.info("수정하기 호출.........................................");
 		log.info(upt);
 		log.info("...........................................");
-		rttr.addFlashAttribute("result", service.modify(upt));
+		if(service.modify(upt) > 0) {
+			rttr.addFlashAttribute("result", service.modify(upt));
+		}
+		rttr.addAttribute("pageNum", cri.getPageNum()); 
+		rttr.addAttribute("amount", cri.getAmount());
 		return "redirect:/culture/list";
 	}
 	
 	@PostMapping("/remove")
-	public String remove(@RequestParam("cno") Long cno, RedirectAttributes rttr) {
+	public String remove(@RequestParam("cno") Long cno, @ModelAttribute("cri") Criteria cri, RedirectAttributes rttr) {
 		log.info("삭제하기 호출...............................................");
 		log.info("cno : " + cno);
 		log.info("...........................................");
-		rttr.addFlashAttribute("result", service.remove(cno)); 
+		if(service.remove(cno) > 0) {
+			rttr.addFlashAttribute("result", service.remove(cno)); 
+		}
+		rttr.addAttribute("pageNum", cri.getPageNum()); 
+		rttr.addAttribute("amount", cri.getAmount());		
 		return "redirect:/culture/list";
 	}
 }
